fix commission

+++ app/Http/Controllers/UserMasterController.php
@@ -20,7 +20,15 @@ class UserMasterController extends Controller
 
     public function import(Request $request)
     {
+        $file = $request->file('file');
+        if (!$file) {
+            return back()->with('error', '❌ กรุณาเลือกไฟล์ก่อนนำเข้า');
+        }
 
+        // ✅ ตรวจสอบเฉพาะ .xlsx
+        if ($file->getClientOriginalExtension() !== 'xlsx') {
+            return back()->with('error', '❌ กรุณาเลือกไฟล์ที่เป็น .xlsx เท่านั้น');
+        }
         UserMaster::truncate();
         Excel::import(new UserMasterImport,request()->file('file'));
